Engage in cooperative locking for the cached SEV certificate
The Enarx CI will exercise SEV code from multiple crates at
(potentially) the same time.

This is problematic because at any given moment one of the tests could
modify the SEV platform's persistent state which would result in other
tests (such as launch tests or integration tests against the SEV
backend) failing.

The 'sev' crate was the first crate to experience the aforementioned
concurrency concerns and they were ameliorated by simply serializing the
tests. This cannot scale (easily) across the crate boundary.

Soon we will have both 'sev' and 'enarx-keepldr' depending on a
consistent view of the platform state during continuous integration
tasks because it is important that the platform state can be cached and
re-used to avoid being ratelimited by AMD's servers when we are
validating the certificate chain.

This patch allows 'sev' to be a friendly SEV-platform neighbor and obey
reader/writer lock semantics with a file lock. Other crates will be
expected to do the same.

This is helpful because for example, running integration tests against
'enarx-keepldr' will result in several SEV loaders spawning. The
reader-writer lock semantics will allow each of these keeps to read from
the cached certificate file without having to wait their turn for the
lock and without fear that the 'sev' dangerous tests could ruin their
launch as a write lock cannot be taken while the read locks have yet to
be released.

Similarly, no reader locks can be taken if a writer lock has yet to be
released.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>