Replace a matches! macro with a logical equivalent
The matches! macro was introduced in Rust 1.42, however, as of
this commit, the Rust build toolchain shipped in CentOS is 1.41,
so the sev crate will fail to build from source. While we do not
currently ship the sev crate as an RPM, it is built as part of
sevctl's dependency graph which I'm working on packaging right
now.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>