feat: run a real SNP VM
The code in the VM does a VM termination request via the GHCB MSR.

In the kernel dmesg buffer, the `5.13.0-0.rc1.14.snp.rfc4.1.fc34.x86_64`
kernel gives us the provides exit code:
```
[323835.755893] SEV-ES guest requested termination: 0x1:0x2
```
but returns the vCPU `run()` with `EINVAL`, which is of course
unfortunate.

This also removes dangerous_hw_tests, as it does no harm.