fix: simplify snp_launch test
Without `vm_fd.create_irq_chip()` the `hlt` instruction causes a
`KVM_EXIT_HLT`, because no IPIs are expected to wakeup the vCPU from
`hlt`. This opens the possibility to reduce the test case to a simple
`hlt` in the VM and test for a proper `Ok(VcpuExit::Hlt)` for the test
to succeed.